refactor(whatsrust): remove final facade-owned definitions
